Fix watch later MD5 sum calculation
This commit will:
- Change the Utility.mpvWatchLaterMd5 method to match the sum calculated
  by mpv
- Add a new ignorePath parameter to mpvWatchLaterMd5 in support of the
  mpv ignore-path-in-watch-later-config option
- Add a new ignorePathInWatchLaterConfig property to PlayerCore that
  gets and caches the value of the mpv ignore-path-in-watch-later-config
  option
- Change PlayerCore.refreshCachedVideoInfo to use the
  ignorePathInWatchLaterConfig property when calling mpvWatchLaterMd5
- Remove the call to mpvWatchLaterMd5 from PlaybackInfo
- Change PlayerCore.openMainWindow to call mpvWatchLaterMd5 and set
  PlaybackInfo.mpvMd5
- Change PlayerCore.fileStarted to call mpvWatchLaterMd5 and set
  PlaybackInfo.mpvMd5
- Add a mpvMd5 parameter to the PlaybackHistory initializer and remove
  the call to mpvWatchLaterMd5
- Add an ignorePath parameter to HistoryController.add
- Change HistoryController.add to call mpvWatchLaterMd5 and pass the sum
  when constructing the PlaybackHistory object
- Change PlayerCore.fileLoaded to pass ignorePathInWatchLaterConfig when
  calling HistoryController.add
- Add a cat_watch_later_file.sh script to other that makes it easy for
  developer to see the contents of a watch later file

This commit will correct the MD5 sum calculation for YouTube videos to
match the sum calculated by mpv. The current code is calculating the sum
based on the URL path, which for a YouTube video is just "/watch". This
prevents IINA from finding the watch later file and also causes the
history file to have only one entry for all YouTube videos.

+++ iina/PlayerSaveState.swift
@@ -806,13 +806,17 @@ struct PlayerSaveState: CustomStringConvertible {
     pwc.osd.lastPlaybackPosition = playbackPositionSec
     pwc.osd.lastPlaybackDuration = playbackDurationSec
 
-    // IINA restore supercedes mpv watch-later.
-    // Need to delete the watch-later file before mpv loads it or else things get very buggy
-    let mpvMD5 = Utility.mpvWatchLaterMd5(url.path)
-    let watchLaterFileURL = Utility.watchLaterURL.appendingPathComponent(mpvMD5).path
-    if FileManager.default.fileExists(atPath: watchLaterFileURL) {
-      player.log.debug("Found mpv watch-later file. Deleting it because we are using IINA restore...")
-      try? FileManager.default.removeItem(atPath: watchLaterFileURL)
+    if player.mpv.getFlag(MPVOption.WatchLater.resumePlayback) {
+      let ignorePathForMD5 = player.mpv.getFlag(MPVOption.WatchLater.ignorePathInWatchLaterConfig)
+      // IINA restore supercedes mpv watch-later.
+      // Need to delete the watch-later file before mpv loads it or else things get very buggy
+      let mpvMD5 = Utility.mpvWatchLaterMd5(url, ignorePathForMD5)
+      let watchLaterFileURL = Utility.watchLaterURL.appendingPathComponent(mpvMD5).path
+      if FileManager.default.fileExists(atPath: watchLaterFileURL) {
+        player.log.debug("Found mpv watch-later file (ignorePathForMD5=\(ignorePathForMD5.yn)). "
+                         + "Deleting it because we are using IINA restore...")
+        try? FileManager.default.removeItem(atPath: watchLaterFileURL)
+      }
     }
 
     if let overrideAutoMusicMode = bool(for: .overrideAutoMusicMode) {
